TruncSilenceBase in lib-builtin-effects
